// shelter-frontend/src/app/api/model/updateFakturyDto.ts
/**
 * Shelter
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateFakturyDto {
    id?: number;
    numer?: string;
    data?: string;
    kwota?: number;
    lecznicaId?: number;
    zwierzeta?: number[];
}